Move screenshotApp and cloneBundle to utils, remove starting '_'
These aren't 'private' to utils.js, so this commit removes the '_' at
the beginning of the function names.
